% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-ROptSpace.R
\docType{package}
\name{ROptSpace}
\alias{ROptSpace}
\alias{package-ROptSpace}
\title{Matrix Reconstruction from a Few Entries}
\description{
Matrix reconstruction, also known as matrix completion, 
is the task of inferring missing entries of a partially observed matrix. 
This package provides a method called OptSpace, 
which was proposed by Keshavan, R.H., Oh, S., and Montanari, A. (2009) 
<doi:10.1109/ISIT.2009.5205567> for a case under low-rank assumption.
}
